PMCVIP-831 - Add a class to the imported image and set it to 100% width
We struggled with which image size to embed in the post, and went over a handful of ideas (e.g. inserting a set image size [but each site has it's own sizes], allowing a default webcam embed size cheezcap[but that's sort of restrictive and widespread], etc.) And landed on simply making the image 100% of the container width.