PMCVIP-831 - Rename the set cookie file back to *.html
I initially renamed this file thinking we'd add some php in there, but this causes VIP to disallow direct access (you can't directly access a *.php file on VIP). Using a *.html file to set the cookie forces VIP to treat this as a static file and then they DO allow direct access.