------------------------------------------------------------------------ r239224 | jamesmehorterpmc | 2016-04-04 17:06:07 +0000 (Mon, 04 Apr 2016) | 3 lines Changed paths:    M /pmc-plugins/webdam-asset-chooser/assets/assetchooser-loader.js    M /pmc-plugins/webdam-asset-chooser/includes/class-asset-chooser.php    M /pmc-plugins/webdam-asset-chooser/includes/set-cookie.html
Manually build the WebDAM set cookie URL

On environments where the admin uses a different domain than the frontend (qa, wpcom, etc.) we need to build the url to this .html file so it uses the admin domain. We'll continue to polish this and try to find a better method of doing so.
------------------------------------------------------------------------