+ Correct an issue where non-admin users are unable to select assets
Previously we were hiding the set-cookie options page with remove_submenu_page in admin_menu, which works for users with manage_options. However, users with lower capabilities are unable to access the hidden page. Move that remove_submenu_page call to admin_head corrects that issue—and all users with edit_posts are now able to visit the hidden page (they're not literally visiting the page, but an AJAX call on their behalf is)

+ Add a user cap check to the webdam hidden set-cookie page — this options page is created for users with edit_posts.. let's ensure only those users can visit this page.